+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/SyntaxErrorCheck.java
@@ -19,6 +19,11 @@ public class SyntaxErrorCheck extends MagikCheck {
   private static final String MESSAGE = "Invalid code.";
 
   @Override
+  public boolean isTemplatedCheck() {
+    return false;
+  }
+
+  @Override
   public List<AstNodeType> subscribedTo() {
     return Arrays.asList(
         MagikGrammar.SYNTAX_ERROR);