+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/checks/UnusedVariableCheck.java
@@ -34,6 +34,11 @@ public class UnusedVariableCheck extends MagikCheck {
   }
 
   @Override
+  public boolean isTemplatedCheck() {
+    return false;
+  }
+
+  @Override
   public List<AstNodeType> subscribedTo() {
     return Arrays.asList(
         MagikGrammar.MAGIK,