+++ magik-lint/src/main/java/org/stevenlooman/sw/magik/lint/CheckInfo.java
@@ -1,5 +1,6 @@
 package org.stevenlooman.sw.magik.lint;
 
+import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.JSONTokener;
 import org.sonar.check.RuleProperty;
@@ -105,6 +106,35 @@ public class CheckInfo {
   public String getSeverity() throws FileNotFoundException {
     JSONObject metadata = readMetadata();
     return metadata.getString("defaultSeverity");
+  }
+
+  /**
+   * Get tags for this check.
+   * @return List of tags.
+   * @throws FileNotFoundException -
+   */
+  public List<String> getTags() throws FileNotFoundException {
+    JSONObject metadata = readMetadata();
+    JSONArray tags = metadata.getJSONArray("tags");
+    List<String> tagsList = new ArrayList<>();
+    for (int i = 0; i < tags.length(); ++i) {
+      String tag = tags.getString(i);
+      tagsList.add(tag);
+    }
+    return tagsList;
+  }
+
+  /**
+   * Get the first tag for this check.
+   * @return First tag.
+   * @throws FileNotFoundException -
+   */
+  public String getTag() throws FileNotFoundException {
+    List<String> tags = getTags();
+    if (tags.isEmpty()) {
+      return null;
+    }
+    return tags.get(0);
   }
 
   public String getTitle() throws FileNotFoundException {